fjern wellknown endepunkt fra IdPortenController, StsController, TokendingsController (fjernes helt) og LoginCookieController
