+++ src/main/kotlin/no/nav/sbl/sosialhjelp_mock_alt/integrations/idporten/model/IdPortenAccessTokenResponse.kt
@@ -9,10 +9,3 @@ data class IdPortenAccessTokenResponse(
     @JsonProperty(value = "expires_in", required = true) val expiresIn: Int,
     @JsonProperty(value = "scope", required = true) val scope: String
 )
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class WellKnown(
-    @JsonProperty(value = "issuer", required = true) val issuer: String,
-    @JsonProperty(value = "token_endpoint", required = true) val tokenEndpoint: String,
-    @JsonProperty(value = "jwks_uri", required = false) val jwksURI: String?
-)
