+++ src/main/kotlin/no/nav/sbl/sosialhjelp_mock_alt/integrations/wellknown/model/WellKnown.kt
@@ -9,3 +9,10 @@ data class WellKnown(
     @JsonProperty(value = "token_endpoint", required = true) val tokenEndpoint: String,
     @JsonProperty(value = "jwks_uri", required = false) val jwksURI: String?
 )
+
+data class TokenResponse(
+    @JsonProperty("access_token") val accessToken: String,
+    @JsonProperty("issued_token_type") val issuedTokenType: String,
+    @JsonProperty("token_type") val tokenType: String,
+    @JsonProperty("expires_in") val expiresIn: Int
+)
